backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic[email]>=2.5.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyJWT>=2.8.0
pymysql>=1.1.0
cryptography>=41.0.0
openai>=1.12.0

